Struct esp_idf_svc::espnow::EspNow

pub struct EspNow<'a>(/* private fields */);

Implementations

impl EspNow<'static>

pub fn take() -> Result<Self, EspError>

impl<'a> EspNow<'a>

pub unsafe fn take_nonstatic() -> Result<Self, EspError>

Safety

This method - in contrast to method take - allows the user to set non-static callbacks/closures into the returned EspNow service. This enables users to borrow

in the closure - variables that live on the stack - or more generally - in the same scope where the service is created.

HOWEVER: care should be taken NOT to call core::mem::forget() on the service, as that would immediately lead to an UB (crash). Also note that forgetting the service might happen with Rc and Arc when circular references are introduced: https://github.com/rust-lang/rust/issues/24456

The reason is that the closure is actually sent to a hidden ESP IDF thread. This means that if the service is forgotten, Rust is free to e.g. unwind the stack and the closure now owned by this other thread will end up with references to variables that no longer exist.

The destructor of the service takes care - prior to the service being dropped and e.g. the stack being unwind - to remove the closure from the hidden thread and destroy it. Unfortunately, when the service is forgotten, the un-subscription does not happen and invalid references are left dangling.

This “local borrowing” will only be possible to express in a safe way once/if !Leak types are introduced to Rust (i.e. the impossibility to “forget” a type and thus not call its destructor).

pub fn send(&self, peer_addr: [u8; 6], data: &[u8]) -> Result<(), EspError>

pub fn add_peer(&self, peer_info: PeerInfo) -> Result<(), EspError>

pub fn del_peer(&self, peer_addr: [u8; 6]) -> Result<(), EspError>

pub fn mod_peer(&self, peer_info: PeerInfo) -> Result<(), EspError>

pub fn get_peer(&self, peer_addr: [u8; 6]) -> Result<PeerInfo, EspError>

pub fn peer_exists(&self, peer_addr: [u8; 6]) -> Result<bool, EspError>

pub fn get_peers_number(&self) -> Result<(usize, usize), EspError>

pub fn fetch_peer(&self, from_head: bool) -> Result<PeerInfo, EspError>

pub fn set_pmk(&self, pmk: &[u8]) -> Result<(), EspError>

pub fn get_version(&self) -> Result<u32, EspError>

pub fn register_recv_cb<F>(&self, callback: F) -> Result<(), EspError>
where F: FnMut(&[u8], &[u8]) + Send + 'a,

pub fn unregister_recv_cb(&self) -> Result<(), EspError>

pub fn register_send_cb<F>(&self, callback: F) -> Result<(), EspError>
where F: FnMut(&[u8], SendStatus) + Send + 'a,

pub fn unregister_send_cb(&self) -> Result<(), EspError>