esp_idf_svc::espnow
Type Alias PeerInfo

pub type PeerInfo = esp_now_peer_info_t;

Aliased Type

struct PeerInfo {
    pub peer_addr: [u8; 6],
    pub lmk: [u8; 16],
    pub channel: u8,
    pub ifidx: u32,
    pub encrypt: bool,
    pub priv_: *mut c_void,
}

Fields
peer_addr: [u8; 6]

< ESPNOW peer MAC address that is also the MAC address of station or softap

lmk: [u8; 16]

< ESPNOW peer local master key that is used to encrypt data

channel: u8

< Wi-Fi channel that peer uses to send/receive ESPNOW data. If the value is 0, use the current channel which station or softap is on. Otherwise, it must be set as the channel that station or softap is on.

ifidx: u32

< Wi-Fi interface that peer uses to send/receive ESPNOW data

encrypt: bool

< ESPNOW data that this peer sends/receives is encrypted or not

priv_: *mut c_void

< ESPNOW peer private data