esp_idf_svc::eventloop
Struct EspEventLoop

pub struct EspEventLoop<T>(/* private fields */)
where
    T: EspEventLoopType;

Implementations

impl<T> EspEventLoop<T>
where T: EspEventLoopType,

pub fn subscribe_async<D>(&self) -> Result<EspAsyncSubscription<D, T>, EspError>
where D: EspEventDeserializer,

pub fn subscribe<D, F>( &self, callback: F, ) -> Result<EspSubscription<'static, T>, EspError>
where D: EspEventDeserializer, F: for<'a> FnMut(D::Data<'a>) + Send + 'static,

pub unsafe fn subscribe_nonstatic<'a, D, F>( &self, callback: F, ) -> Result<EspSubscription<'a, T>, EspError>
where D: EspEventDeserializer, F: for<'d> FnMut(D::Data<'d>) + Send + 'a,

Safety

This method - in contrast to method subscribe - allows the user to pass a non-static callback/closure. This enables users to borrow

in the closure - variables that live on the stack - or more generally - in the same scope where the service is created.

HOWEVER: care should be taken NOT to call core::mem::forget() on the service, as that would immediately lead to an UB (crash). Also note that forgetting the service might happen with Rc and Arc when circular references are introduced: https://github.com/rust-lang/rust/issues/24456

The reason is that the closure is actually sent to a hidden ESP IDF thread. This means that if the service is forgotten, Rust is free to e.g. unwind the stack and the closure now owned by this other thread will end up with references to variables that no longer exist.

The destructor of the service takes care - prior to the service being dropped and e.g. the stack being unwind - to remove the closure from the hidden thread and destroy it. Unfortunately, when the service is forgotten, the un-subscription does not happen and invalid references are left dangling.

This โlocal borrowingโ will only be possible to express in a safe way once/if !Leak types are introduced to Rust (i.e. the impossibility to โforgetโ a type and thus not call its destructor).

pub async fn post_async<S>(&self, payload: &S::Data<'_>) -> Result<(), EspError>
where S: EspEventSerializer,

pub fn post<S>( &self, payload: &S::Data<'_>, timeout: TickType_t, ) -> Result<bool, EspError>
where S: EspEventSerializer,

impl<T> EspEventLoop<User<T>>

pub fn spin(&mut self, timeout: TickType_t) -> Result<(), EspError>

impl EspEventLoop<System>

pub fn take() -> Result<Self, EspError>

impl EspEventLoop<User<Background>>

pub fn new(conf: &BackgroundLoopConfiguration<'_>) -> Result<Self, EspError>

impl EspEventLoop<User<Explicit>>

pub fn new(conf: &ExplicitLoopConfiguration) -> Result<Self, EspError>