esp_idf_svc::eventloop
Struct EspEventPostData

pub struct EspEventPostData<'a> { /* private fields */ }

Implementations

impl<'a> EspEventPostData<'a>

pub unsafe fn new<P: Copy + Send + 'static>( source: &'static CStr, event_id: Option<i32>, payload: &'a P, ) -> Self

Safety

Care should be taken to only call this function with payload reference that lives at least as long as the call that will post this data to the event loop

pub unsafe fn new_raw( source: &'static CStr, event_id: Option<i32>, payload: &'a [u8], ) -> Self

Safety

Care should be taken to only call this function with payload reference that lives at least as long as the call that will post this data to the event loop