Module esp_idf_svc::hal::prelude

The prelude.

To use the esp_idf_hal effectively, a lot of traits and types need to be imported. Instead of importing them one by one manually, the prelude contains the most commonly used imports that are used around application runtime management.

This can be imported as use esp_idf_hal::prelude::*.

Structs

Hertz

HertzU64

KiloHertz

KiloHertzU64

MegaHertz

MegaHertzU64

MicroSeconds

MicroSecondsU64

MilliSeconds

MilliSecondsU64

NanoSeconds

NanoSecondsU64

Peripherals

Seconds

SecondsU64

Ticks

TicksU64

Traits

Count

CountU64

Frequency

FrequencyU64

FromLargeValueType

FromValueType

Quantity

Time

TimeU64

Type Aliases

LargeValueType

ValueType