Function esp_idf_svc::hal::sys::adc_continuous_register_event_callbacks

pub unsafe extern "C" fn adc_continuous_register_event_callbacks(
    handle: *mut adc_continuous_ctx_t,
    cbs: *const adc_continuous_evt_cbs_t,
    user_data: *mut c_void
) -> i32

@brief Register callbacks

@note User can deregister a previously registered callback by calling this function and setting the to-be-deregistered callback member in the cbs structure to NULL. @note When CONFIG_ADC_CONTINUOUS_ISR_IRAM_SAFE is enabled, the callback itself and functions called by it should be placed in IRAM. Involved variables (including user_data) should be in internal RAM as well. @note You should only call this API when the ADC continuous mode driver isn’t started. Check return value to know this.

@param[in] handle ADC continuous mode driver handle @param[in] cbs Group of callback functions @param[in] user_data User data, which will be delivered to the callback functions directly

@return - ESP_OK: On success - ESP_ERR_INVALID_ARG: Invalid arguments - ESP_ERR_INVALID_STATE: Driver state is invalid, you shouldn’t call this API at this moment