esp_idf_svc::hal::sys

Function ble_gap_ext_disc

pub unsafe extern "C" fn ble_gap_ext_disc(
    own_addr_type: u8,
    duration: u16,
    period: u16,
    filter_duplicates: u8,
    filter_policy: u8,
    limited: u8,
    uncoded_params: *const ble_gap_ext_disc_params,
    coded_params: *const ble_gap_ext_disc_params,
    cb: Option<unsafe extern "C" fn(*mut ble_gap_event, *mut c_void) -> i32>,
    cb_arg: *mut c_void,
) -> i32

Performs the Limited or General Extended Discovery Procedures.

@param own_addr_type The type of address the stack should use for itself when sending scan requests. Valid values are: - BLE_ADDR_TYPE_PUBLIC - BLE_ADDR_TYPE_RANDOM - BLE_ADDR_TYPE_RPA_PUB_DEFAULT - BLE_ADDR_TYPE_RPA_RND_DEFAULT This parameter is ignored unless active scanning is being used. @param duration The duration of the discovery procedure. On expiration, if period is set to 0, the procedure ends and a BLE_GAP_EVENT_DISC_COMPLETE event is reported. Units are 10 milliseconds. Specify 0 for no expiration. @param period Time interval from when the Controller started its last Scan Duration until it begins the subsequent Scan Duration. Specify 0 to scan continuously. Units are 1.28 second. @param filter_duplicates Set to enable packet filtering in the controller @param filter_policy Set the used filter policy. Valid values are: - BLE_HCI_SCAN_FILT_NO_WL - BLE_HCI_SCAN_FILT_USE_WL - BLE_HCI_SCAN_FILT_NO_WL_INITA - BLE_HCI_SCAN_FILT_USE_WL_INITA - BLE_HCI_SCAN_FILT_MAX This parameter is ignored unless @p filter_duplicates is set. @param limited If limited discovery procedure should be used. @param uncoded_params Additional arguments specifying the particulars of the discovery procedure for uncoded PHY. If NULL is provided no scan is performed for this PHY. @param coded_params Additional arguments specifying the particulars of the discovery procedure for coded PHY. If NULL is provided no scan is performed for this PHY. @param cb The callback to associate with this discovery procedure. Advertising reports and discovery termination events are reported through this callback. @param cb_arg The optional argument to pass to the callback function.

@return 0 on success; nonzero on failure.