esp_idf_svc::hal::sys

Function ble_gatts_notify_multiple_custom

pub unsafe extern "C" fn ble_gatts_notify_multiple_custom(
    conn_handle: u16,
    chr_count: usize,
    tuples: *mut ble_gatt_notif,
) -> i32

Sends a “free-form” multiple handle variable length characteristic notification. This function consumes supplied mbufs regardless of the outcome. Notifications are sent in order of supplied entries. Function tries to send minimum amount of PDUs. If PDU can’t contain all of the characteristic values, multiple notifications are sent. If only one handle-value pair fits into PDU, or only one characteristic remains in the list, regular characteristic notification is sent.

If GATT client doesn’t support receiving multiple handle notifications, this will use GATT notification for each characteristic, separately.

If value of characteristic is not specified it will be read from local GATT database.

@param conn_handle The connection over which to execute the procedure. @param chr_count Number of characteristics to notify about. @param tuples Handle-value pairs in form of ble_gatt_notif structures.

@return 0 on success; nonzero on failure.