esp_idf_svc::hal::sys
Function otLogGenerateNextHexDumpLine

pub unsafe extern "C" fn otLogGenerateNextHexDumpLine(
    aInfo: *mut otLogHexDumpInfo,
) -> u32

Generates the next hex dump line.

Can call this method back-to-back to generate the hex dump output line by line. On the first call the mIterator field in @p aInfo MUST be set to zero.

Here is an example of the generated hex dump output:

“==========================[{mTitle} len=070]============================” “| 41 D8 87 34 12 FF FF 25 | 4C 57 DA F2 FB 2F 62 7F | A..4…%LW…/b. |” “| 3B 01 F0 4D 4C 4D 4C 54 | 4F 00 15 15 00 00 00 00 | ;..MLMLTO……. |” “| 00 00 00 01 80 DB 60 82 | 7E 33 72 3B CC B3 A1 84 | ……`.~3r;…. |” “| 3B E6 AD B2 0B 45 E7 45 | C5 B9 00 1A CB 2D 6D 1C | ;….E.E…..-m. |” “| 10 3E 3C F5 D3 70 | | .><..p |” “————————————————————————”

@param[in,out] aInfo A pointer to otLogHexDumpInfo to use to generate hex dump.

@retval OT_ERROR_NONE Successfully generated the next line, mLine field in @p aInfo is updated. @retval OT_ERROR_NOT_FOUND Reached the end and no more line to generate.