Function esp_idf_svc::hal::sys::psa_sign_hash_get_num_ops

pub unsafe extern "C" fn psa_sign_hash_get_num_ops(
    operation: *const psa_sign_hash_interruptible_operation_s
) -> u32

\brief Get the number of ops that a hash signing operation has taken so far. If the operation has completed, then this will represent the number of ops required for the entire operation. After initialization or calling \c psa_sign_hash_interruptible_abort() on the operation, a value of 0 will be returned.

\note This interface is guaranteed re-entrant and thus may be called from driver code.

\warning This is a beta API, and thus subject to change at any point. It is not bound by the usual interface stability promises.

                         This is a helper provided to help you tune the
                         value passed to \c
                         psa_interruptible_set_max_ops().

\param operation The \c psa_sign_hash_interruptible_operation_t to use. This must be initialized first.

\return Number of ops that the operation has taken so far.