Struct esp_idf_svc::hal::uart::UartTxDriver

pub struct UartTxDriver<'d> { /* private fields */ }

Serial transmitter

Implementations

impl<'d> UartTxDriver<'d>

pub fn new<UART>( uart: impl Peripheral<P = UART> + 'd, tx: impl Peripheral<P = impl OutputPin> + 'd, cts: Option<impl Peripheral<P = impl InputPin> + 'd>, rts: Option<impl Peripheral<P = impl OutputPin> + 'd>, config: &Config ) -> Result<UartTxDriver<'d>, EspError>
where UART: Uart,

Create a new serial transmitter

pub fn event_queue(&self) -> Option<&Queue<UartEvent>>

Retrieves the event queue for this UART. Returns None if the config specified 0 for queue_size.

pub fn change_stop_bits( &self, stop_bits: StopBits ) -> Result<&UartTxDriver<'d>, EspError>

Change the number of stop bits

pub fn stop_bits(&self) -> Result<StopBits, EspError>

Returns the current number of stop bits

pub fn change_data_bits( &self, data_bits: DataBits ) -> Result<&UartTxDriver<'d>, EspError>

Change the number of data bits

pub fn data_bits(&self) -> Result<DataBits, EspError>

Return the current number of data bits

pub fn change_parity( &self, parity: Parity ) -> Result<&UartTxDriver<'d>, EspError>

Change the type of parity checking

pub fn parity(&self) -> Result<Parity, EspError>

Returns the current type of parity checking

pub fn change_baudrate<T>( &self, baudrate: T ) -> Result<&UartTxDriver<'d>, EspError>
where T: Into<Hertz> + Copy,

Change the baudrate.

Will automatically select the clock source. When possible the reference clock (1MHz) will be used, because this is constant when the clock source/frequency changes. However if one of the clock frequencies is below 10MHz or if the baudrate is above the reference clock or if the baudrate cannot be set within 1.5% then use the APB clock.

pub fn baudrate(&self) -> Result<Hertz, EspError>

Returns the current baudrate

pub fn write(&mut self, bytes: &[u8]) -> Result<usize, EspError>

Write multiple bytes from a slice

pub fn write_nb(&self, bytes: &[u8]) -> Result<usize, EspError>

Write multiple bytes from a slice directly to the TX FIFO hardware. Returns the number of bytes written, where 0 would mean that the TX FIFO is full.

NOTE: In case the UART TX buffer is enabled, this method might have unpredictable results when used together with method write, as the latter will push the data to be sent to the TX buffer first.

To avoid this, always call wait_done after the last call to write and before calling this method.

pub fn wait_done(&self, timeout: u32) -> Result<(), EspError>

Waits until the transmission is complete or until the specified timeout expires.

pub fn flush(&mut self) -> Result<(), EspError>

👎Deprecated since 0.41.3: Use UartTxDriver::wait_done instead

Waits until the transmission is complete.

pub fn port(&self) -> i32

pub fn count(&self) -> Result<usize, EspError>

Get count of remaining capacity in the transmit ring buffer