esp_idf_svc::handle

Trait RawHandle

pub trait RawHandle {
    type Handle;

    // Required method
    fn handle(&self) -> Self::Handle;
}

Required Associated Types

type Handle

Required Methods

fn handle(&self) -> Self::Handle

Care should be taken to use the returned ESP-IDF driver raw handle only while the driver is still alive, so as to avoid use-after-free errors.

Dyn Compatibility

This trait is dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety".