esp_idf_svc::http::server
Struct EspHttpConnection

pub struct EspHttpConnection<'a> { /* private fields */ }

Implementations

impl<'a> EspHttpConnection<'a>

Represents the two-way connection between an HTTP request and its response.

pub fn uri(&self) -> &str

pub fn method(&self) -> Method

pub fn header(&self, name: &str) -> Option<&str>

pub fn split(&mut self) -> (&EspHttpConnection<'a>, &mut Self)

pub fn initiate_response( &mut self, status: u16, message: Option<&str>, headers: &[(&str, &str)], ) -> Result<(), EspError>

Sends the HTTP status (e.g. “200 OK”) and the response headers to the HTTP client.

pub fn is_response_initiated(&self) -> bool

Returns true if the response headers have been sent to the HTTP client.

pub fn read(&mut self, buf: &mut [u8]) -> Result<usize, EspError>

Reads bytes from the body of the HTTP request.

This is typically used whenever the HTTP server has to parse the body of an HTTP POST request.

server.fn_handler("/foo", Method::Post, move |mut request| {
    let (_headers, connection) = request.split();
    let mut buffer: [u8; 1024] = [0; 1024];
    let bytes_read = connection.read(&mut buffer)?;

    let my_data = MyDataStruct::from_bytes(&buffer[0..bytes_read]);
    // etc

pub fn write(&mut self, buf: &[u8]) -> Result<usize, EspError>

Sends bytes back to the HTTP client; returns the number of bytes sent.

pub fn write_all(&mut self, buf: &[u8]) -> Result<(), EspError>

pub fn raw_connection( &mut self, ) -> Result<&mut EspHttpRawConnection<'a>, EspError>