esp_idf_svc::http::server

Trait EspHttpTraversableChainNonstatic

pub unsafe trait EspHttpTraversableChainNonstatic<'a>: EspHttpTraversableChain<'a> { }

Safety

Implementing this trait means that the chain can contain non-'static handlers and that the chain can be used with method EspHttpServer::handler_chain_nonstatic.

Consult the documentation of EspHttpServer::handler_chain_nonstatic for more information on how to use non-static handler chains.

Dyn Compatibility

This trait is dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety".