Enum esp_idf_svc::http::server::ws::EspHttpWsConnection

pub enum EspHttpWsConnection {
    New(httpd_handle_t, *mut httpd_req_t),
    Receiving(httpd_handle_t, *mut httpd_req_t, Option<httpd_ws_frame_t>),
    Closed(c_int),
}

A Websocket connection between this server and a client.

Variants

New(httpd_handle_t, *mut httpd_req_t)

Receiving(httpd_handle_t, *mut httpd_req_t, Option<httpd_ws_frame_t>)

Closed(c_int)

Implementations

impl EspHttpWsConnection

pub fn session(&self) -> i32

pub fn is_new(&self) -> bool

Returns true when the connection still hasn’t received any data

pub fn is_closed(&self) -> bool

Returns true when the connection already has been closed.

pub fn create_detached_sender( &self, ) -> Result<EspHttpWsDetachedSender, EspError>

pub fn send( &mut self, frame_type: FrameType, frame_data: &[u8], ) -> Result<(), EspError>

Sends a frame to the client.

pub fn recv( &mut self, frame_data_buf: &mut [u8], ) -> Result<(FrameType, usize), EspError>

Receives a frame from the client.