esp_idf_svc::ipv4
Struct SocketAddrV4

pub struct SocketAddrV4 { /* private fields */ }

An IPv4 socket address.

IPv4 socket addresses consist of an IPv4 address and a 16-bit port number, as stated in IETF RFC 793.

See SocketAddr for a type encompassing both IPv4 and IPv6 socket addresses.

The size of a SocketAddrV4 struct may vary depending on the target operating system. Do not assume that this type has the same memory layout as the underlying system representation.

Textual representation

SocketAddrV4 provides a FromStr implementation. It accepts an IPv4 address in its textual representation, followed by a single :, followed by the port encoded as a decimal integer. Other formats are not accepted.

Examples

use std::net::{Ipv4Addr, SocketAddrV4};

let socket = SocketAddrV4::new(Ipv4Addr::new(127, 0, 0, 1), 8080);

assert_eq!("127.0.0.1:8080".parse(), Ok(socket));
assert_eq!(socket.ip(), &Ipv4Addr::new(127, 0, 0, 1));
assert_eq!(socket.port(), 8080);

Implementations

impl SocketAddrV4

pub fn parse_ascii(b: &[u8]) -> Result<SocketAddrV4, AddrParseError>

๐ฌThis is a nightly-only experimental API. (addr_parse_ascii)

Parse an IPv4 socket address from a slice of bytes.

#![feature(addr_parse_ascii)]

use std::net::{Ipv4Addr, SocketAddrV4};

let socket = SocketAddrV4::new(Ipv4Addr::new(127, 0, 0, 1), 8080);

assert_eq!(SocketAddrV4::parse_ascii(b"127.0.0.1:8080"), Ok(socket));

impl SocketAddrV4

pub const fn new(ip: Ipv4Addr, port: u16) -> SocketAddrV4

Creates a new socket address from an IPv4 address and a port number.

Examples

use std::net::{SocketAddrV4, Ipv4Addr};

let socket = SocketAddrV4::new(Ipv4Addr::new(127, 0, 0, 1), 8080);

pub const fn ip(&self) -> &Ipv4Addr

Returns the IP address associated with this socket address.

Examples

use std::net::{SocketAddrV4, Ipv4Addr};

let socket = SocketAddrV4::new(Ipv4Addr::new(127, 0, 0, 1), 8080);
assert_eq!(socket.ip(), &Ipv4Addr::new(127, 0, 0, 1));

pub fn set_ip(&mut self, new_ip: Ipv4Addr)

Changes the IP address associated with this socket address.

Examples

use std::net::{SocketAddrV4, Ipv4Addr};

let mut socket = SocketAddrV4::new(Ipv4Addr::new(127, 0, 0, 1), 8080);
socket.set_ip(Ipv4Addr::new(192, 168, 0, 1));
assert_eq!(socket.ip(), &Ipv4Addr::new(192, 168, 0, 1));

pub const fn port(&self) -> u16

Returns the port number associated with this socket address.

Examples

use std::net::{SocketAddrV4, Ipv4Addr};

let socket = SocketAddrV4::new(Ipv4Addr::new(127, 0, 0, 1), 8080);
assert_eq!(socket.port(), 8080);

pub fn set_port(&mut self, new_port: u16)

Changes the port number associated with this socket address.

Examples

use std::net::{SocketAddrV4, Ipv4Addr};

let mut socket = SocketAddrV4::new(Ipv4Addr::new(127, 0, 0, 1), 8080);
socket.set_port(4242);
assert_eq!(socket.port(), 4242);