esp_idf_svc::log

Trait LogFilterBackend

pub trait LogFilterBackend {
    // Required method
    fn enabled(&self, metadata: &Metadata<'_>) -> bool;

    // Provided method
    fn initialize(&self) { ... }
}

Trait for a log filter backend that can be used with the EspIdfLogger.

Required Methods

fn enabled(&self, metadata: &Metadata<'_>) -> bool

Check if logging for the given metadata is enabled.

Provided Methods

fn initialize(&self)

Initialize the log filter backend.

Dyn Compatibility

This trait is dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety".