esp_idf_svc::mqtt::client
Struct EspAsyncMqttClient

pub struct EspAsyncMqttClient(/* private fields */);

Implementations

impl EspAsyncMqttClient

pub fn new( url: &str, conf: &MqttClientConfiguration<'_>, ) -> Result<(Self, EspAsyncMqttConnection), EspError>

Create a new MQTT client with a given URL and configuration.

pub fn new_with_caps( url: &str, conf: &MqttClientConfiguration<'_>, caps: Option<(usize, usize, usize)>, ) -> Result<(Self, EspAsyncMqttConnection), EspError>

Create a new MQTT client with a given URL, configuration and caps.

The caps tuple contains the initial capacity of the topic, payload and broker URI buffers. Useful to avoid constant re-allocations with large payloads.

pub fn wrap(client: EspMqttClient<'static>) -> Result<Self, EspError>

Wrap an existing MQTT client with an async interface.

pub fn wrap_with_caps( client: EspMqttClient<'static>, caps: Option<(usize, usize, usize)>, ) -> Result<Self, EspError>

Wrap an existing MQTT client with an async interface.

The caps tuple contains the initial capacity of the topic, payload and broker URI buffers. Useful to avoid constant re-allocations with large payloads.

pub async fn subscribe( &mut self, topic: &str, qos: QoS, ) -> Result<MessageId, EspError>

pub async fn unsubscribe(&mut self, topic: &str) -> Result<MessageId, EspError>

pub async fn publish( &mut self, topic: &str, qos: QoS, retain: bool, payload: &[u8], ) -> Result<MessageId, EspError>

pub async fn set_uri(&mut self, broker_uri: &str) -> Result<MessageId, EspError>