esp_idf_svc::ota

Struct EspOtaUpdate

pub struct EspOtaUpdate<'a> { /* private fields */ }

Implementations

impl<'a> EspOtaUpdate<'a>

pub fn write(&mut self, buf: &[u8]) -> Result<(), EspError>

Writes OTA update data to partition. This function can be called multiple times as data is received during the OTA operation. Data is written sequentially to the partition.

Errors

Returns an error if data could not be written to flash.

pub fn flush(&mut self) -> Result<(), EspError>

This function does not perform any flash operations, as flash writes are not cached and, therefore, do not need to be flushed.

Errors

Returns an error update partition is not valid.

pub fn finish(self) -> Result<EspOtaUpdateFinished<'a>, EspError>

Finishes the OTA update and validates the new app image. Returns an instance of EspOtaUpdateFinished.

This function does not update the boot partition. The user must call activate() on the returned instance of EspOtaUpdateFinished.

See also: complete

pub fn complete(self) -> Result<(), EspError>

Completes the OTA process by validating the new app image and updating the boot partition.

pub fn abort(self) -> Result<(), EspError>

Cancels the update.