esp_idf_svc

Module partition

ESP IDF partitions API

This API provides access to the partitions in the ESP32 flash memory - with operations for reading, writing, and erasing. The ESP-IDF Wear-Leveling algorithm is also supported.

Note that ESP-IDF partitions are not created or dropped by this API - they always pre-existing and the API provides access to them. To define your partitions, you need to use the ESP-IDF partition table CSV file, as described here: https://docs.espressif.com/projects/esp-idf/en/latest/esp32/api-guides/partition-tables.html

Structs

EspMemMappedPartition

Represents a memory-mapping of a partition region

EspPartition

Represents a partition in the ESP32 flash memory

EspPartitionIterator

An iterator over the partitions in the ESP32 flash memory

EspWlPartition

Represents a partition wrapped with the ESP-IDF Wear-Leveling algorithm

Enums

EspAppPartitionSubtype

The subtype of an application partition

EspDataPartitionSubtype

The subtype of a data partition

EspMemMapType

The type of memory mapping

EspPartitionType

The type of a partition