Function esp_idf_svc::sys::i2s_channel_init_pdm_tx_mode

pub unsafe extern "C" fn i2s_channel_init_pdm_tx_mode(
    handle: *mut i2s_channel_obj_t,
    pdm_tx_cfg: *const i2s_pdm_tx_config_t
) -> i32

@brief Initialize i2s channel to PDM TX mode @note Only allowed to be called when the channel state is REGISTERED, (i.e., channel has been allocated, but not initialized) and the state will be updated to READY if initialization success, otherwise the state will return to REGISTERED.

@param[in] handle I2S tx channel handler @param[in] pdm_tx_cfg Configurations for PDM TX mode, including clock, slot and gpio The clock configuration can be generated by the helper macro I2S_PDM_TX_CLK_DEFAULT_CONFIG The slot configuration can be generated by the helper macro I2S_PDM_TX_SLOT_DEFAULT_CONFIG

@return - ESP_OK Initialize successfully - ESP_ERR_NO_MEM No memory for storing the channel information - ESP_ERR_INVALID_ARG NULL pointer or invalid configuration - ESP_ERR_INVALID_STATE This channel is not registered