esp_idf_svc::sys
Function otInstanceGetUptimeAsString

pub unsafe extern "C" fn otInstanceGetUptimeAsString(
    aInstance: *mut otInstance,
    aBuffer: *mut u8,
    aSize: u16,
)

Returns the current instance uptime as a human-readable string.

Requires OPENTHREAD_CONFIG_UPTIME_ENABLE to be enabled.

The string follows the format “::.” for hours, minutes, seconds and millisecond (if uptime is shorter than one day) or “
d.::.” (if longer than a day).

If the resulting string does not fit in @p aBuffer (within its @p aSize characters), the string will be truncated but the outputted string is always null-terminated.

@param[in] aInstance A pointer to an OpenThread instance. @param[out] aBuffer A pointer to a char array to output the string. @param[in] aSize The size of @p aBuffer (in bytes). Recommended to use OT_UPTIME_STRING_SIZE.