esp_idf_svc::sys

Function otIp6UnsubscribeMulticastAddress

pub unsafe extern "C" fn otIp6UnsubscribeMulticastAddress(
    aInstance: *mut otInstance,
    aAddress: *const otIp6Address,
) -> u32

Unsubscribes the Thread interface to a Network Interface Multicast Address.

@param[in] aInstance A pointer to an OpenThread instance. @param[in] aAddress A pointer to an IP Address.

@retval OT_ERROR_NONE Successfully unsubscribed to the Network Interface Multicast Address. @retval OT_ERROR_REJECTED The IP Address indicated by @p aAddress is an internal address. @retval OT_ERROR_NOT_FOUND The IP Address indicated by @p aAddress was not found.