Struct esp_idf_svc::sys::esp_bt_controller_config_t

#[repr(C)]
pub struct esp_bt_controller_config_t {

    pub magic: u32,
    pub version: u32,
    pub controller_task_stack_size: u16,
    pub controller_task_prio: u8,
    pub controller_task_run_cpu: u8,
    pub bluetooth_mode: u8,
    pub ble_max_act: u8,
    pub sleep_mode: u8,
    pub sleep_clock: u8,
    pub ble_st_acl_tx_buf_nb: u8,
    pub ble_hw_cca_check: u8,
    pub ble_adv_dup_filt_max: u16,
    pub coex_param_en: bool,
    pub ce_len_type: u8,
    pub coex_use_hooks: bool,
    pub hci_tl_type: u8,
    pub hci_tl_funcs: *mut esp_bt_hci_tl_t,
    pub txant_dft: u8,
    pub rxant_dft: u8,
    pub txpwr_dft: u8,
    pub cfg_mask: u32,
    pub scan_duplicate_mode: u8,
    pub scan_duplicate_type: u8,
    pub normal_adv_size: u16,
    pub mesh_adv_size: u16,
    pub coex_phy_coded_tx_rx_time_limit: u8,
    pub hw_target_code: u32,
    pub slave_ce_len_min: u8,
    pub hw_recorrect_en: u8,
    pub cca_thresh: u8,
    pub scan_backoff_upperlimitmax: u16,
    pub dup_list_refresh_period: u16,
    pub ble_50_feat_supp: bool,
    pub ble_cca_mode: u8,

}

@brief Controller config options, depend on config mask. Config mask indicate which functions enabled, this means some options or parameters of some functions enabled by config mask.

Fields
magic: u32

< Magic number

version: u32

< version number of the defined structure

controller_task_stack_size: u16

< Bluetooth controller task stack size

controller_task_prio: u8

< Bluetooth controller task priority

controller_task_run_cpu: u8

< CPU num that Bluetooth controller task runs on

bluetooth_mode: u8

< Controller mode: BR/EDR, BLE or Dual Mode

ble_max_act: u8

< BLE maximum number of air activities

sleep_mode: u8

< controller sleep mode

sleep_clock: u8

< controller sleep clock

ble_st_acl_tx_buf_nb: u8

< controller static ACL TX BUFFER number

ble_hw_cca_check: u8

< controller hardware triggered CCA check

ble_adv_dup_filt_max: u16

< maxinum number of duplicate scan filter

coex_param_en: bool

< deprecated

ce_len_type: u8

< connection event length computation method

coex_use_hooks: bool

< deprecated

hci_tl_type: u8

< HCI transport layer, UART, VHCI, etc

hci_tl_funcs: *mut esp_bt_hci_tl_t

< hci transport functions used, must be set when hci_tl_type is UART

txant_dft: u8

< default Tx antenna

rxant_dft: u8

< default Rx antenna

txpwr_dft: u8

< default Tx power

cfg_mask: u32

< Configuration mask to set specific options

scan_duplicate_mode: u8

< scan duplicate mode

scan_duplicate_type: u8

< scan duplicate type

normal_adv_size: u16

< Normal adv size for scan duplicate

mesh_adv_size: u16

< Mesh adv size for scan duplicate

coex_phy_coded_tx_rx_time_limit: u8

< limit on max tx/rx time in case of connection using CODED-PHY with Wi-Fi coexistence

hw_target_code: u32

< hardware target

slave_ce_len_min: u8

< slave minimum ce length

hw_recorrect_en: u8

< Hardware re-correction enabled

cca_thresh: u8

< cca threshold

scan_backoff_upperlimitmax: u16

< scan backoff upperlimitmax value

dup_list_refresh_period: u16

< duplicate scan list refresh time

ble_50_feat_supp: bool

< BLE 5.0 feature support

ble_cca_mode: u8

< BLE CCA mode