Type Alias esp_idf_svc::sys::TaskSnapshot_t

pub type TaskSnapshot_t = xTASK_SNAPSHOT;

@brief Task Snapshot structure

Used with the uxTaskGetSnapshotAll() function to save memory snapshot of each task in the system.

We need this structure because TCB_t is defined (hidden) in tasks.c.

Aliased Type

struct TaskSnapshot_t {
    pub pxTCB: *mut c_void,
    pub pxTopOfStack: *mut u8,
    pub pxEndOfStack: *mut u8,
}

Fields
pxTCB: *mut c_void

< Address of the task control block.

pxTopOfStack: *mut u8

< Points to the location of the last item placed on the tasks stack.

pxEndOfStack: *mut u8

< Points to the end of the stack. pxTopOfStack < pxEndOfStack, stack grows hi2lo pxTopOfStack > pxEndOfStack, stack grows lo2hi