Type Alias esp_idf_svc::sys::esp_pm_config_esp32s3_t

pub type esp_pm_config_esp32s3_t = esp_pm_config_t;

@brief Power management config

Pass a pointer to this structure as an argument to esp_pm_configure function.

Aliased Type

struct esp_pm_config_esp32s3_t {
    pub max_freq_mhz: i32,
    pub min_freq_mhz: i32,
    pub light_sleep_enable: bool,
}

Fields
max_freq_mhz: i32

< Maximum CPU frequency, in MHz

min_freq_mhz: i32

< Minimum CPU frequency to use when no locks are taken, in MHz

light_sleep_enable: bool

< Enter light sleep when no locks are taken