Type Alias esp_idf_svc::sys::esp_vfs_fat_sdmmc_mount_config_t

pub type esp_vfs_fat_sdmmc_mount_config_t = esp_vfs_fat_mount_config_t;

@brief Configuration arguments for esp_vfs_fat_sdmmc_mount and esp_vfs_fat_spiflash_mount_rw_wl functions

Aliased Type

struct esp_vfs_fat_sdmmc_mount_config_t {
    pub format_if_mount_failed: bool,
    pub max_files: i32,
    pub allocation_unit_size: usize,
    pub disk_status_check_enable: bool,
}

Fields
format_if_mount_failed: bool

If FAT partition can not be mounted, and this parameter is true, create partition table and format the filesystem.

max_files: i32

< Max number of open files

allocation_unit_size: usize

If format_if_mount_failed is set, and mount fails, format the card with given allocation unit size. Must be a power of 2, between sector size and 128 * sector size. For SD cards, sector size is always 512 bytes. For wear_levelling, sector size is determined by CONFIG_WL_SECTOR_SIZE option.

Using larger allocation unit size will result in higher read/write performance and higher overhead when storing small files.

Setting this field to 0 will result in allocation unit set to the sector size.

disk_status_check_enable: bool

Enables real ff_disk_status function implementation for SD cards (ff_sdmmc_status). Possibly slows down IO performance.

Try to enable if you need to handle situations when SD cards are not unmounted properly before physical removal or you are experiencing issues with SD cards.

Doesn’t do anything for other memory storage media.