esp_idf_svc::thread

Enum ThreadEvent

pub enum ThreadEvent {

    Started,
    Stopped,
    Detached,
    Attached,
    RoleChanged {
        current_role: Role,
        previous_role: Role,
    },
    IfUp,
    IfDown,
    GotIpv6,
    LostIpv6,
    MulticastJoined,
    MulticastLeft,
    TrelIpv6Added,
    TrelIpv6Removed,
    TrelMulticastJoined,
    DnsServerSet,
    MeshcopEPublishStarted,
    MeshcopERemoveStarted,
    DatasetChanged,
    Other(i32),

}

Events reported by the Thread stack on the system event loop

Variants

Started

Thread stack started

Stopped

Thread stack stopped

Detached

Thread stack detached

Attached

Thread stack attached

RoleChanged

Thread role changed

Fields

current_role: Role

previous_role: Role

IfUp

Thread network interface up

IfDown

Thread network interface down

GotIpv6

Thread got IPv6 address

LostIpv6

Thread lost IPv6 address

MulticastJoined

Thread multicast group joined

MulticastLeft

Thread multicast group left

TrelIpv6Added

Thread TREL IPv6 address added

TrelIpv6Removed

Thread TREL IPv6 address removed

TrelMulticastJoined

Thread TREL multicast group joined

DnsServerSet

Thread DNS server set

MeshcopEPublishStarted

Thread Meshcop E Publish started

MeshcopERemoveStarted

Thread Meshcop E Remove started

DatasetChanged

Other(i32)

An event ID not recognised by this version of the library was received.

This variant is produced instead of panicking when an unknown event ID arrives, allowing applications to remain forward-compatible with ESP-IDF versions that introduce new Thread events.