esp_idf_svc::thread

Struct SrpService

pub struct SrpService<'a, SI, TI> {
    pub name: &'a str,
    pub instance_name: &'a str,
    pub subtype_labels: SI,
    pub txt_entries: TI,
    pub port: u16,
    pub priority: u16,
    pub weight: u16,
    pub lease_secs: u32,
    pub key_lease_secs: u32,
}

An SRP service that can be registered with the OpenThread stack.

Fields
name: &'a str

The service name.

instance_name: &'a str

The instance name.

subtype_labels: SI

The subtype labels.

txt_entries: TI

The TXT entries.

port: u16

The service port.

priority: u16

The service priority.

weight: u16

The service weight.

lease_secs: u32

The service lease time in seconds. Set to 0 to use the default value as specified in SrpConf.

key_lease_secs: u32

The service key lease time in seconds. Set to 0 to use the default value as specified in SrpConf.