esp_idf_svc::thread

Type Alias OutSrpService

pub type OutSrpService<'a> = SrpService<'a, OutSrpSubtypeLabelsIter<'a>, OutSrpTxtEntriesIter<'a>>;

Type alias for an SRP service as returned by OpenThread::srp_services.

Aliased Type

pub struct OutSrpService<'a> {
    pub name: &'a str,
    pub instance_name: &'a str,
    pub subtype_labels: OutSrpSubtypeLabelsIter<'a>,
    pub txt_entries: OutSrpTxtEntriesIter<'a>,
    pub port: u16,
    pub priority: u16,
    pub weight: u16,
    pub lease_secs: u32,
    pub key_lease_secs: u32,
}

Fields
name: &'a str

The service name.

instance_name: &'a str

The instance name.

subtype_labels: OutSrpSubtypeLabelsIter<'a>

The subtype labels.

txt_entries: OutSrpTxtEntriesIter<'a>

The TXT entries.

port: u16

The service port.

priority: u16

The service priority.

weight: u16

The service weight.

lease_secs: u32

The service lease time in seconds. Set to 0 to use the default value as specified in SrpConf.

key_lease_secs: u32

The service key lease time in seconds. Set to 0 to use the default value as specified in SrpConf.