Struct esp_idf_svc::tls::EspTls

pub struct EspTls<S>
where
    S: Socket,
{ /* private fields */ }

Wrapper for esp-tls module. Only supports synchronous operation for now.

Implementations

impl EspTls<InternalSocket>

pub fn new() -> Result<Self, EspError>

Create a new EspTls instance using internally-managed socket.

Errors

ESP_ERR_NO_MEM if not enough memory to create the TLS connection

pub fn connect( &mut self, host: &str, port: u16, cfg: &Config<'_>, ) -> Result<CompletedHandshake, EspError>

Establish a TLS/SSL connection with the specified host and port, using an internally-managed socket.

Errors

ESP_ERR_INVALID_SIZE if cfg.alpn_protos exceeds 9 elements or avg 10 bytes/ALPN

ESP_FAIL if connection could not be established

ESP_TLS_ERR_SSL_WANT_READ if the socket is in non-blocking mode and it is not ready for reading

ESP_TLS_ERR_SSL_WANT_WRITE if the socket is in non-blocking mode and it is not ready for writing

EWOULDBLOCK if the socket is in non-blocking mode and it is not ready either for reading or writing (a peculiarity/bug of the esp-tls C module)

impl<S> EspTls<S>
where S: Socket,

pub fn adopt(socket: S) -> Result<Self, EspError>

Create a new EspTls instance adopting the supplied socket. The socket should be in a connected state.

Errors

ESP_ERR_NO_MEM if not enough memory to create the TLS connection

pub fn negotiate( &mut self, host: &str, cfg: &Config<'_>, ) -> Result<CompletedHandshake, EspError>

Establish a TLS/SSL connection using the adopted socket.

Errors

ESP_ERR_INVALID_SIZE if cfg.alpn_protos exceeds 9 elements or avg 10 bytes/ALPN

ESP_FAIL if connection could not be established

ESP_TLS_ERR_SSL_WANT_READ if the socket is in non-blocking mode and it is not ready for reading

ESP_TLS_ERR_SSL_WANT_WRITE if the socket is in non-blocking mode and it is not ready for writing

EWOULDBLOCK if the socket is in non-blocking mode and it is not ready either for reading or writing (a peculiarity/bug of the esp-tls C module)

pub fn read(&mut self, buf: &mut [u8]) -> Result<usize, EspError>

Read in the supplied buffer. Returns the number of bytes read.

Errors

ESP_TLS_ERR_SSL_WANT_READ if the socket is in non-blocking mode and it is not ready for reading

ESP_TLS_ERR_SSL_WANT_WRITE if the socket is in non-blocking mode and it is not ready for writing

Any other EspError for a general error

pub fn write(&mut self, buf: &[u8]) -> Result<usize, EspError>

Write the supplied buffer. Returns the number of bytes written.

Errors

ESP_TLS_ERR_SSL_WANT_READ if the socket is in non-blocking mode and it is not ready for reading

ESP_TLS_ERR_SSL_WANT_WRITE if the socket is in non-blocking mode and it is not ready for writing

Any other EspError for a general error

pub fn write_all(&mut self, buf: &[u8]) -> Result<(), EspError>

pub fn context_handle(&self) -> *mut esp_tls