esp_idf_svc

Module wifi

WiFi support

Modules

config

Structs

AccessPointConfiguration

AccessPointInfo

ApStaConnectedRef

ApStaDisconnectedRef

ApWrongPasswordRef

Payload reference for WifiEvent::ApWrongPassword.

AsyncWifi

Wraps a WifiDriver or EspWifi, and offers strictly async (non-blocking) function calls for their functionality.

BlockingWifi

Wraps a WifiDriver or EspWifi, and offers strictly synchronous (blocking) function calls for their functionality.

ClientConfiguration

Configuration for wifi in STA mode

DppCfgRecvdRef

Payload reference for WifiEvent::DppCfgRecvd.

DppFailedRef

Payload reference for WifiEvent::DppFailed.

DppUriReadyRef

Payload reference for WifiEvent::DppUriReady.

EspWifi

EspWifi wraps a WifiDriver Data Link layer instance, and binds the OSI Layer 3 (network) facilities of ESP IDF to it.

HomeChannelChange

StaBeaconOffsetUnstableRef

Payload reference for WifiEvent::StaBeaconOffsetUnstable.

StaConnectedRef

StaDisconnectedRef

StaNeighborRepRef

Payload reference for WifiEvent::StaNeighborRep.

StaScanDoneRef

WifiDriver

This struct provides a safe wrapper over the ESP IDF Wifi C driver.

WifiFrame

WpsConfig

WpsCredentials

WpsCredentialsRef

WpsFactoryInfo

Enums

AuthMethod

Capability

Configuration

PmfConfiguration

Protected Management Frame configuration

Protocol

ScanMethod

The scan method to use when connecting to an AP

ScanSortMethod

SecondaryChannel

WifiDeviceId

WifiEvent

WpsStatus

WpsType

Traits

NonBlocking