Crate esp_idf_svc

This crate contains wrappers which are mostly implementations of the abstractions defined in the embedded-svc project. It has features such as wifi, networking, httpd, logging.

Features

This crate specifies a few cargo features, including:

std: Enable the use of std. Enabled by default.

experimental: Enable the use of experimental features.

embassy-time-driver: Implement an embassy time driver.

Modules

espnow

ESP-NOW

eth

eventloop

Event loop library

hal

handle

Wrapper around ESP-IDF raw handles

http

HTTP server and client

io

ipv4

log

Logging

mqtt

MQTT protocol

netif

Network abstraction

nvs

Non-Volatile Storage (NVS)

ota

Over The Air Updates (OTA)

ping

Send ICMP echo requests (Ping)

sntp

SNTP Time Synchronization

sys

systime

System time

timer

High resolution hardware timer based task scheduling

tls

Type safe abstraction for esp-tls

wifi

WiFi support

ws

WebSocket protocol