esp_idf_svc::handle
Trait RawHandle

pub trait RawHandle {
    type Handle;

    // Required method
    fn handle(&self) -> Self::Handle;
}

Required Associated Types

type Handle

Required Methods

fn handle(&self) -> Self::Handle

Care should be taken to use the returned ESP-IDF driver raw handle only while the driver is still alive, so as to avoid use-after-free errors.