Struct esp_idf_svc::mqtt::client::EspMqttClient

pub struct EspMqttClient<'a, S = ()> { /* private fields */ }

Implementations

impl EspMqttClient<'static, AsyncConnState<MessageImpl, EspError>>

pub fn new_with_async_conn( url: &str, conf: &MqttClientConfiguration<'_> ) -> Result<(Self, EspMqttAsyncConnection), EspError>where Self: Sized,

impl<M, E> EspMqttClient<'static, AsyncConnState<M, E>>where M: Send + 'static, E: Debug + Send + 'static,

pub fn new_with_converting_async_conn<F>( url: &str, conf: &MqttClientConfiguration<'_>, converter: F ) -> Result<(Self, EspMqttConvertingAsyncConnection<M, E>), EspError>where F: for<'b> FnMut(&'b Result<Event<EspMqttMessage<'b>>, EspError>) -> Result<Event<M>, E> + Send + 'static, Self: Sized,

impl<'a, M, E> EspMqttClient<'a, AsyncConnState<M, E>>where M: Send + 'static, E: Debug + Send + 'static,

pub unsafe fn new_nonstatic_with_converting_async_conn<F>( url: &str, conf: &MqttClientConfiguration<'_>, converter: F ) -> Result<(Self, EspMqttConvertingAsyncConnection<M, E>), EspError>where F: for<'b> FnMut(&'b Result<Event<EspMqttMessage<'b>>, EspError>) -> Result<Event<M>, E> + Send + 'a, Self: Sized,

Safety

This method - in contrast to method new_with_converting_async_conn - allows the user to pass a non-static callback/closure. This enables users to borrow

in the closure - variables that live on the stack - or more generally - in the same scope where the service is created.

HOWEVER: care should be taken NOT to call core::mem::forget() on the service, as that would immediately lead to an UB (crash). Also note that forgetting the service might happen with Rc and Arc when circular references are introduced: https://github.com/rust-lang/rust/issues/24456

The reason is that the closure is actually sent to a hidden ESP IDF thread. This means that if the service is forgotten, Rust is free to e.g. unwind the stack and the closure now owned by this other thread will end up with references to variables that no longer exist.

The destructor of the service takes care - prior to the service being dropped and e.g. the stack being unwind - to remove the closure from the hidden thread and destroy it. Unfortunately, when the service is forgotten, the un-subscription does not happen and invalid references are left dangling.

This “local borrowing” will only be possible to express in a safe way once/if !Leak types are introduced to Rust (i.e. the impossibility to “forget” a type and thus not call its destructor).

impl EspMqttClient<'static, ConnState<MessageImpl, EspError>>

pub fn new_with_conn( url: &str, conf: &MqttClientConfiguration<'_> ) -> Result<(Self, Connection<RawCondvar, MessageImpl, EspError>), EspError>where Self: Sized,

impl<M, E> EspMqttClient<'static, ConnState<M, E>>where M: Send + 'static, E: Debug + Send + 'static,

pub fn new_with_converting_conn<F>( url: &str, conf: &MqttClientConfiguration<'_>, converter: F ) -> Result<(Self, Connection<RawCondvar, M, E>), EspError>where F: for<'b> FnMut(&'b Result<Event<EspMqttMessage<'b>>, EspError>) -> Result<Event<M>, E> + Send + 'static, Self: Sized,

impl<'a, M, E> EspMqttClient<'a, ConnState<M, E>>where M: Send + 'static, E: Debug + Send + 'static,

pub unsafe fn new_nonstatic_with_converting_conn<F>( url: &str, conf: &MqttClientConfiguration<'_>, converter: F ) -> Result<(Self, Connection<RawCondvar, M, E>), EspError>where F: for<'b> FnMut(&'b Result<Event<EspMqttMessage<'b>>, EspError>) -> Result<Event<M>, E> + Send + 'a, Self: Sized,

Safety

This method - in contrast to method new_with_converting_conn - allows the user to pass a non-static callback/closure. This enables users to borrow

in the closure - variables that live on the stack - or more generally - in the same scope where the service is created.

HOWEVER: care should be taken NOT to call core::mem::forget() on the service, as that would immediately lead to an UB (crash). Also note that forgetting the service might happen with Rc and Arc when circular references are introduced: https://github.com/rust-lang/rust/issues/24456

The reason is that the closure is actually sent to a hidden ESP IDF thread. This means that if the service is forgotten, Rust is free to e.g. unwind the stack and the closure now owned by this other thread will end up with references to variables that no longer exist.

The destructor of the service takes care - prior to the service being dropped and e.g. the stack being unwind - to remove the closure from the hidden thread and destroy it. Unfortunately, when the service is forgotten, the un-subscription does not happen and invalid references are left dangling.

This “local borrowing” will only be possible to express in a safe way once/if !Leak types are introduced to Rust (i.e. the impossibility to “forget” a type and thus not call its destructor).

impl EspMqttClient<'static, ()>

pub fn new<F>( url: &str, conf: &MqttClientConfiguration<'_>, callback: F ) -> Result<Self, EspError>where F: for<'b> FnMut(&'b Result<Event<EspMqttMessage<'b>>, EspError>) + Send + 'static, Self: Sized,

impl<'a> EspMqttClient<'a, ()>

pub unsafe fn new_nonstatic<F>( url: &str, conf: &MqttClientConfiguration<'_>, callback: F ) -> Result<Self, EspError>where F: for<'b> FnMut(&'b Result<Event<EspMqttMessage<'b>>, EspError>) + Send + 'a, Self: Sized,

Safety

This method - in contrast to method new - allows the user to pass a non-static callback/closure. This enables users to borrow

in the closure - variables that live on the stack - or more generally - in the same scope where the service is created.

HOWEVER: care should be taken NOT to call core::mem::forget() on the service, as that would immediately lead to an UB (crash). Also note that forgetting the service might happen with Rc and Arc when circular references are introduced: https://github.com/rust-lang/rust/issues/24456

The reason is that the closure is actually sent to a hidden ESP IDF thread. This means that if the service is forgotten, Rust is free to e.g. unwind the stack and the closure now owned by this other thread will end up with references to variables that no longer exist.

The destructor of the service takes care - prior to the service being dropped and e.g. the stack being unwind - to remove the closure from the hidden thread and destroy it. Unfortunately, when the service is forgotten, the un-subscription does not happen and invalid references are left dangling.

This “local borrowing” will only be possible to express in a safe way once/if !Leak types are introduced to Rust (i.e. the impossibility to “forget” a type and thus not call its destructor).

impl<S> EspMqttClient<'static, S>

pub fn new_generic<F>( url: &str, conf: &MqttClientConfiguration<'_>, conn_state_guard: Option<Arc<ConnStateGuard<RawCondvar, S>>>, callback: F ) -> Result<Self, EspError>where F: for<'b> FnMut(&'b Result<Event<EspMqttMessage<'b>>, EspError>) + Send + 'static, Self: Sized,

impl<'a, S> EspMqttClient<'a, S>

pub unsafe fn new_nonstatic_generic<F>( url: &str, conf: &MqttClientConfiguration<'_>, conn_state_guard: Option<Arc<ConnStateGuard<RawCondvar, S>>>, callback: F ) -> Result<Self, EspError>where F: for<'b> FnMut(&'b Result<Event<EspMqttMessage<'b>>, EspError>) + Send + 'a, Self: Sized,

Safety

This method - in contrast to method new_generic - allows the user to pass a non-static callback/closure. This enables users to borrow

in the closure - variables that live on the stack - or more generally - in the same scope where the service is created.

HOWEVER: care should be taken NOT to call core::mem::forget() on the service, as that would immediately lead to an UB (crash). Also note that forgetting the service might happen with Rc and Arc when circular references are introduced: https://github.com/rust-lang/rust/issues/24456

The reason is that the closure is actually sent to a hidden ESP IDF thread. This means that if the service is forgotten, Rust is free to e.g. unwind the stack and the closure now owned by this other thread will end up with references to variables that no longer exist.

The destructor of the service takes care - prior to the service being dropped and e.g. the stack being unwind - to remove the closure from the hidden thread and destroy it. Unfortunately, when the service is forgotten, the un-subscription does not happen and invalid references are left dangling.

This “local borrowing” will only be possible to express in a safe way once/if !Leak types are introduced to Rust (i.e. the impossibility to “forget” a type and thus not call its destructor).

pub fn subscribe( &mut self, topic: &str, qos: QoS ) -> Result<MessageId, EspError>

pub fn unsubscribe(&mut self, topic: &str) -> Result<MessageId, EspError>

pub fn publish( &mut self, topic: &str, qos: QoS, retain: bool, payload: &[u8] ) -> Result<MessageId, EspError>

pub fn enqueue( &mut self, topic: &str, qos: QoS, retain: bool, payload: &[u8] ) -> Result<MessageId, EspError>