Module esp_idf_svc::wifi

WiFi support

Modules

config

Structs

AccessPointConfiguration

AccessPointInfo

ApStaConnectedRef

ApStaDisconnectedRef

AsyncWifi

Wraps a WifiDriver or EspWifi, and offers strictly async (non-blocking) function calls for their functionality.

BlockingWifi

Wraps a WifiDriver or EspWifi, and offers strictly synchronous (blocking) function calls for their functionality.

ClientConfiguration

Configuration for wifi in STA mode

EspWifi

EspWifi wraps a WifiDriver Data Link layer instance, and binds the OSI Layer 3 (network) facilities of ESP IDF to it. In other words, it connects the ESP IDF AP and STA Netif interfaces to the Wifi driver. This allows users to utilize the Rust STD APIs for working with TCP and UDP sockets.

WifiDriver

This struct provides a safe wrapper over the ESP IDF Wifi C driver. The driver works on Layer 2 (Data Link) in the OSI model, in that it provides facilities for sending and receiving ethernet packets over the WiFi radio.

WifiFrame

WpsConfig

WpsCredentials

WpsCredentialsRef

WpsFactoryInfo

Enums

AuthMethod

Capability

Configuration

PmfConfiguration

Protected Management Frame configuration

Protocol

ScanMethod

The scan method to use when connecting to an AP

ScanSortMethod

SecondaryChannel

WifiDeviceId

WifiEvent

WpsStatus

WpsType

Traits

NonBlocking