Struct esp_idf_svc::wifi::EspWifi

pub struct EspWifi<'d> { /* private fields */ }

EspWifi wraps a WifiDriver Data Link layer instance, and binds the OSI Layer 3 (network) facilities of ESP IDF to it. In other words, it connects the ESP IDF AP and STA Netif interfaces to the Wifi driver. This allows users to utilize the Rust STD APIs for working with TCP and UDP sockets.

This struct should be the default option for a Wifi driver in all use cases but the niche one where bypassing the ESP IDF Netif and lwIP stacks is desirable. E.g., using smoltcp or other custom IP stacks on top of the ESP IDF Wifi radio.

Implementations

impl<'d> EspWifi<'d>

pub fn new<M: WifiModemPeripheral>( modem: impl Peripheral<P = M> + 'd, sysloop: EspSystemEventLoop, nvs: Option<EspDefaultNvsPartition> ) -> Result<Self, EspError>

pub fn wrap(driver: WifiDriver<'d>) -> Result<Self, EspError>

pub fn wrap_all( driver: WifiDriver<'d>, sta_netif: EspNetif, ap_netif: EspNetif ) -> Result<Self, EspError>

pub fn swap_netif( &mut self, sta_netif: EspNetif, ap_netif: EspNetif ) -> Result<(EspNetif, EspNetif), EspError>

Replaces the network interfaces with the given ones. Returns the old ones.

pub fn driver(&self) -> &WifiDriver<'d>

Returns the underlying WifiDriver

pub fn driver_mut(&mut self) -> &mut WifiDriver<'d>

Returns the underlying WifiDriver, as mutable

pub fn sta_netif(&self) -> &EspNetif

Returns the underlying EspNetif for client mode

pub fn sta_netif_mut(&mut self) -> &mut EspNetif

Returns the underlying EspNetif for client mode, as mutable

pub fn ap_netif(&self) -> &EspNetif

Returns the underlying EspNetif for AP mode

pub fn ap_netif_mut(&mut self) -> &mut EspNetif

Returns the underlying EspNetif for AP mode, as mutable

pub fn get_capabilities(&self) -> Result<EnumSet<Capability>, EspError>

As per WifiDriver::get_capabilities()

pub fn is_started(&self) -> Result<bool, EspError>

As per WifiDriver::is_started()

pub fn is_connected(&self) -> Result<bool, EspError>

As per WifiDriver::is_connected()

pub fn is_up(&self) -> Result<bool, EspError>

Returns true when the driver has a connection, it has enabled either client or AP mode, and either the client or AP network interface is up.

pub fn get_configuration(&self) -> Result<Configuration, EspError>

As per WifiDriver::get_configuration()

pub fn set_configuration( &mut self, conf: &Configuration ) -> Result<(), EspError>

As per WifiDriver::set_configuration()

pub fn start(&mut self) -> Result<(), EspError>

As per WifiDriver::start()

pub fn stop(&mut self) -> Result<(), EspError>

As per WifiDriver::stop()

pub fn connect(&mut self) -> Result<(), EspError>

As per WifiDriver::connect()

pub fn disconnect(&mut self) -> Result<(), EspError>

As per WifiDriver::disconnect()

pub fn is_scan_done(&self) -> Result<bool, EspError>

As per WifiDriver::is_scan_done()

pub fn scan_n<const N: usize>( &mut self ) -> Result<(Vec<AccessPointInfo, N>, usize), EspError>

As per WifiDriver::scan_n()

pub fn scan(&mut self) -> Result<Vec<AccessPointInfo>, EspError>

As per WifiDriver::scan()

pub fn start_scan( &mut self, scan_config: &ScanConfig, blocking: bool ) -> Result<(), EspError>

As per WifiDriver::start_scan().

pub fn stop_scan(&mut self) -> Result<(), EspError>

As per WifiDriver::stop_scan().

pub fn get_scan_result_n<const N: usize>( &mut self ) -> Result<(Vec<AccessPointInfo, N>, usize), EspError>

As per WifiDriver::get_scan_result_n().

pub fn get_scan_result(&mut self) -> Result<Vec<AccessPointInfo>, EspError>

As per WifiDriver::get_scan_result().

pub fn get_mac(&self, interface: WifiDeviceId) -> Result<[u8; 6], EspError>

As per WifiDriver::get_mac().

pub fn set_mac( &mut self, interface: WifiDeviceId, mac: [u8; 6] ) -> Result<(), EspError>

As per WifiDriver::set_mac().